functions for table_writer and table_reader in _utilities
